requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.7.2
    # via django
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.0
    # via requests
django==3.2
    # via -r requirements.in
django-heroku==0.3.1
idna==3.4
    # via requests
psycopg2-binary==2.9.9
    # via -r requirements.in
pytz==2023.3.post1
    # via django
requests==2.31.0
    # via -r requirements.in
sqlparse==0.4.4
    # via django
urllib3==2.0.6
    # via requests
gunicorn==21.2.0  # Replace with the desired version

djangorestframework==3.14.0

graphene-django==3.1.5

lxml==4.9.3

django-cors-headers==4.3.0